% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ratings.R
\name{get_apple_ratings}
\alias{get_apple_ratings}
\title{Apple App Store Ratings}
\usage{
get_apple_ratings(id, country)
}
\arguments{
\item{id}{The ID of the App on the Apple Store}

\item{country}{The two-letter country code for the store you want to search.
For a list of country codes see \url{http://en.wikipedia.org/wiki/\%20ISO_3166-1_alpha-2}}
}
\value{
A list containing:
\describe{
\item{\code{ovr_rating}}{The overall rating of the App}
\item{\code{n_ratings}}{The number of ratings for the App}
\item{\code{proportion}}{The percentage split of ratings}
}
}
\description{
Scrapes the app page to get the number of ratings, the average rating and the split between
1 and 5 stars
}
