% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reviews.R
\name{get_apple_reviews}
\alias{get_apple_reviews}
\title{Apple Store Reviews}
\usage{
get_apple_reviews(id, country = "us")
}
\arguments{
\item{id}{The ID of the iTunes entity}

\item{country}{The two-letter country code for the store you want to search.
For a list of country codes see \url{http://en.wikipedia.org/wiki/\%20ISO_3166-1_alpha-2}}
}
\value{
A \code{data.frame} of any results that match the iTunes database.

If there were no successful results then it will return \code{NULL}.
}
\description{
Using Apple's iTunes API, will extract the information about a selected
}
\examples{
# Look up Jack Johnson by iTunes artist ID
lookup_apple(909253)

# Look up Jack Johnson by AMG artist ID
lookup_apple(468749, "amgArtistId")

}
\seealso{
\url{https://affiliate.itunes.apple.com/resources/documentation/itunes-store-web-service-search-api/}
}
